Resolved merge conflicts and renamed the utility class.
