--- src/main/java/edu/pav/PatientTrackerSystem/commons/utils.java
@@ -1,33 +1,0 @@
-package edu.pav.PatientTrackerSystem.commons;
-
-import java.security.MessageDigest;
-import java.security.NoSuchAlgorithmException;
-
-public class utils {
-    public static String encryptPassword(String password) {
-        try
-        {
-            /* MessageDigest instance for MD5. */
-            MessageDigest m = MessageDigest.getInstance("MD5");
-
-            /* Add plain-text password bytes to digest using MD5 update() method. */
-            m.update(password.getBytes());
-
-            /* Convert the hash value into bytes */
-            byte[] bytes = m.digest();
-
-            /* The bytes array has bytes in decimal form. Converting it into hexadecimal format. */
-            StringBuilder s = new StringBuilder();
-            for (byte aByte : bytes) {
-                s.append(Integer.toString((aByte & 0xff) + 0x100, 16).substring(1));
-            }
-
-            /* Complete hashed password in hexadecimal format */
-            return s.toString();
-        } catch (NoSuchAlgorithmException e) {
-            e.printStackTrace();
-        }
-
-        return "";
-    }
-}
